requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
openai==1.58.1
torch
torchvision
pillow==11.0.0
python-multipart==0.0.19
jinja2==3.1.4
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
pydantic==2.10.3
pydantic-settings==2.7.0
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
greenlet
boto3